local-client/requirements.txt
aiohttp>=3.9.0
websockets>=12.0
psutil>=5.9.0
pyyaml>=6.0
python-dotenv>=1.0.0